Add WPF support
* Add WpfAutoMeasurement based on WinFormAutoMeasurement
* Delete WinForms project
* Add both WinForms and WPF implementations to the .NET 4.5 project, so
they are both available in the .NET 4.5 nuget
* Refactor AutoMeasurement to a static singleton with public
getter/setter of implementation
* Separate AutoMeasurement to its own file

Notes:
* .NET 4.5 nuget can no longer assume WinForms will be used. This
means when adding the nuget, one must manually add a one-liner that
assigns either WinFormAutoMeasurement or WpfAutoMeasurement to the
AutoMeasurement.Instance property. The readme would have to be updated
to reflect this.

* WpfAutoMeasurement tries to report info about screen
where the main window is currently located, and falls back to primary
screen, whereas WinForms always uses primary screen.