Not relying on exception for program flow
If I run our application with the library and would like to catch all
exceptions not only those unhandled; then the Restore method will always
throw an exception (if there id no file). With this change it will check
if it exists and then return the default if it didnt exist. There is no
need to try to read the stream from the file because it will not be
there.