Make it possible to change Folder path
I would like to be able to set the base folder path to our
application's local app data folder. Otherwise the files will not be
removed when our application is removed.  Since the method is public but
the class is only internal, I made the class public.

I know that we discussed that the default implementation could use the
product name, etc. But this doesnt help because my code is loaded by a
third party application, so that means that the files will be named like
the application and not base on our codebase.